src/main/resources/python/requirements.txt
absl-py==2.0.0
annotated-types==0.6.0
anyio==4.2.0
astunparse==1.6.3
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
cymem==2.0.8
fastapi==0.108.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.2
gast==0.5.4
gensim==4.3.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
langcodes==3.3.0
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
preshed==3.0.9
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
pyfiglet==1.0.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
starlette==0.32.0.post1
summarization==0.0.1
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.4.0
thinc==8.2.2
threadpoolctl==3.2.0
tokenizers==0.15.0
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.36.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.14.1
googlesearch-python